Add Baremetal Topology Panel
In the dashboard, if you click on Network > Network Topology, it
shows a graph topology of the networks and instances. This panel
does not work for baremetal nodes which ESI uses. This adds a
topology view of the baremetal nodes and their MAC address
connections to network ports in addition to the already existing
network topology.