Reorganize Code Structure
Previously, the panels were Project -> ESI -> nodes. Now,
It goes ESI -> Lessee -> nodes, with plans to have Lessee,
Owner, and Admin groups under ESI. The logic from
esi_api.node_list was changed for performance and to lump in
as much info as it could, and then just have the javascript
filter what is being shown the the user. This is all in
preparation to add an Owner's view to manage lease information.