Check if token is project scoped; if not, create a project scoped token
Using password auth results in a project scoped token; however an OIDC
token does not. An unscoped token does not have an attached service catalog,
resulting in errors with the ESI connection code. Checking if a token is
unscoped, and creating a project scoped token if so, fixes the issue.